+++ docs/changelog.rst
@@ -25,6 +25,7 @@ Changed
 - Updated to hydromt v1 API and dropped support for Python 3.10 (PR #162)
 - Updated to hydromt 1.4.0 (PR #254 and PR #302)
 - when reading a model prefer crs from mesh/network file, then geoms, then userdefined (PR #271)
+- Allow writing 'days since' to bc forcing file as time unit (PR #276)
 
 Fixed
 -----